ai-medi-bot/requirements.txt
fastapi
uvicorn
jinja2
transformers
torch
pillow
openai
python-dotenv
python-multipart
google-generativeai


